requirements.txt
web3
pandas
python-dotenv
google.cloud